refactor(module-common): Exception, Message DatabaseClenup 등 공통 내용 모듈 분리
