+++ module-common/src/main/java/kitchenpos/common/exception/InvalidPriceException.java
@@ -1,6 +1,6 @@
-package kitchenpos.exception;
+package kitchenpos.common.exception;
 
-import static kitchenpos.common.ErrorMessage.INVALID_PRICE;
+import static kitchenpos.common.message.ErrorMessage.*;
 
 public class InvalidPriceException extends IllegalArgumentException {
