+++ module-common/src/main/java/kitchenpos/common/exception/InvalidTableNumberException.java
@@ -1,6 +1,6 @@
-package kitchenpos.exception;
+package kitchenpos.common.exception;
 
-import static kitchenpos.common.ErrorMessage.INVALID_TABLE_NUMBER;
+import static kitchenpos.common.message.ErrorMessage.*;
 
 public class InvalidTableNumberException extends IllegalArgumentException {
     public InvalidTableNumberException() {
