+++ module-common/src/main/resources/db/migration/V3__Decmial_to_bigint_change.sql
@@ -1,0 +1,2 @@
+ALTER TABLE menu ALTER COLUMN price TYPE BIGINT(20);
+ALTER TABLE product ALTER COLUMN price TYPE BIGINT(20);
