+++ module-common/src/test/java/kitchenpos/common/ModuleCommonApplicationTests.java
@@ -1,0 +1,13 @@
+package kitchenpos.common;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+class ModuleCommonApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
